Develop

+++ src/main/java/com/zamazor/market/modules/product/service/CategoryService.java
@@ -1,7 +1,7 @@
 package com.zamazor.market.modules.product.service;
 
 import com.zamazor.market.modules.product.exception.CategoryNotFoundException;
-import com.zamazor.market.modules.product.models.dto.CategoryAlreadyExistsException;
+import com.zamazor.market.modules.product.exception.CategoryAlreadyExistsException;
 import com.zamazor.market.modules.product.models.dto.CategoryDto;
 import com.zamazor.market.modules.product.models.dto.CategoryRequest;
 import com.zamazor.market.modules.product.models.entity.Category;
@@ -21,7 +21,7 @@ public class CategoryService {
 
 	public CategoryDto create(CategoryRequest request) {
 		if (categoryRepository.existsByLabel(request.label())) {
-			throw new CategoryAlreadyExistsException("Category with label " + request.label() + " already exists");
+			throw new CategoryAlreadyExistsException("Category with label %s already exists".formatted(request.label()));
 		}
 		var category = new Category();
 		category.setLabel(request.label());
@@ -34,10 +34,10 @@ public class CategoryService {
 
 	public CategoryDto update(UUID id, CategoryRequest request) {
 		var category = categoryRepository.findById(id)
-				.orElseThrow(() -> new CategoryNotFoundException("Category with id: " + id + " not found"));
+				.orElseThrow(() -> new CategoryNotFoundException(id));
 
 		if (categoryRepository.existsByLabel(request.label())) {
-			throw new CategoryAlreadyExistsException("Category with label " + request.label() + " already exists");
+			throw new CategoryAlreadyExistsException("Category with label %s already exists".formatted(request.label()));
 		}
 
 		category.setLabel(request.label());
@@ -45,6 +45,9 @@ public class CategoryService {
 	}
 
 	public void delete(UUID id) {
+		if (!categoryRepository.existsById(id)) {
+			throw new CategoryNotFoundException(id);
+		}
 		categoryRepository.deleteById(id);
 	}
 }
