+++ src/main/java/com/zamazor/market/modules/product/service/ProductService.java
@@ -25,6 +25,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -41,7 +42,7 @@ public class ProductService {
 	public ProductDto create(CreateProductRequest request, @NonNull MultipartFile image) throws IOException {
 		var product = productMapper.toEntity(request);
 		var category = categoryRepository.findById(request.categoryId())
-				.orElseThrow(() -> new CategoryNotFoundException("Category with id: " + request.categoryId() + " not found"));
+				.orElseThrow(() -> new CategoryNotFoundException(request.categoryId()));
 		var store = storeRepository.findOne()
 				.orElseThrow(() -> new IllegalStateException("The single store instance is missing from the database!"));
 
@@ -54,15 +55,9 @@ public class ProductService {
 		return productMapper.toDto(productRepository.save(product));
 	}
 
-	public PageResponse<ProductDto> getAll(Pageable pageable) {
-		Page<ProductDto> productPage = productRepository
-				.findAll(pageable).map(productMapper::toDto);
-		return new PageResponse<>(productPage);
-	}
-
 	public PageResponse<ProductDto> getByCategory(UUID categoryId, Pageable pageable) {
 		if (!categoryRepository.existsById(categoryId)) {
-			throw new CategoryNotFoundException("Category with id: " + categoryId + " not found");
+			throw new CategoryNotFoundException(categoryId);
 		}
 
 		Page<ProductDto> productPage = productRepository
@@ -84,22 +79,29 @@ public class ProductService {
 		return new PageResponse<>(productPage);
 	}
 
+	public PageResponse<ProductDto> bulk(List<UUID> productIds, Pageable pageable) {
+		Page<ProductDto> productPage = productRepository.findByIdIn(productIds, pageable)
+				.map(productMapper::toDto);
+
+		return new PageResponse<>(productPage);
+	}
+
 	public ProductDto getById(UUID id) {
 		return productRepository.findById(id)
 				.map(productMapper::toDto)
-				.orElseThrow(() -> new ProductNotFoundException("Product with id: " + id + " not found"));
+				.orElseThrow(() -> new ProductNotFoundException(id));
 	}
 
 	@Transactional
 	public ProductDto update(UUID id, @NonNull UpdateProductRequest request, MultipartFile image) throws IOException {
 		var product = productRepository.findById(id)
-				.orElseThrow(() -> new ProductNotFoundException("Product with id: " + id + " not found"));
+				.orElseThrow(() -> new ProductNotFoundException(id));
 
 		productMapper.update(request, product);
 
 		if (request.categoryId() != null && !request.categoryId().equals(product.getCategory().getId())) {
 			var category = categoryRepository.findById(request.categoryId())
-					.orElseThrow(() -> new CategoryNotFoundException("Category with id: " + request.categoryId() + " not found"));
+					.orElseThrow(() -> new CategoryNotFoundException(request.categoryId()));
 			product.setCategory(category);
 		}
 		if (image != null && image.getSize() > 0) {
@@ -111,7 +113,11 @@ public class ProductService {
 		return productMapper.toDto(product);
 	}
 
+	@Transactional
 	public void delete(UUID id) {
+		if (!productRepository.existsById(id)) {
+			throw new ProductNotFoundException(id);
+		}
 		productRepository.deleteById(id);
 	}
 }
